requirements.txt
fastapi==0.115.0
pdfplumber==0.11.4
protobuf==5.28.2
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.4
requests==2.32.3
tenacity==9.0.0
uvicorn==0.31.0
